fix: moved dto to database
