+++ apps/prisma-postgres/src/main.ts
@@ -1,6 +1,6 @@
 import bodyParser from "body-parser";
 import express from "express";
-import { createPrismaConnection } from "./connection/connection";
+import { createPrismaConnection } from "./database/connection";
 import restRoutes from "./routes";
 
 try {
