+++ apps/prisma-postgres/tsconfig.app.json
@@ -6,5 +6,5 @@
     "types": ["node", "express"]
   },
   "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts", "src/connection/connection.ts"]
+  "include": ["src/**/*.ts", "src/database/connection.ts"]
 }
